Initial push
Updateing an old project. Fns are from the origional project, but I am starting with the new df (population) which is needed to reweigh the numbers used in the old project. I will update fn as I go and reintegrate the rest of the old project to this new file