Update chi_greenspace_mortality.py
decided to put a pause on makeing merge better.
added reset_index to parse_death (not sure how this worked in the past without this line!)
removed the drop.na line from merge, now that I'm merging in pop this drops most cols. Need to see if I was dropping NAs to make the df easyer to look at or if there was a functional reason
added back in cov_check and ols